[FEATURE] Add compatibility to TYPO3 10 for unit tests
